# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Viagens",
        "description": "API para gerenciamento de solicitações de viagem",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento local"
        }
    ],
    "paths": {
...
        "/api/trip-requests": {
            "get": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Listar todas as solicitações de viagem",
                "operationId": "3d3c332a8f3103b087218d724870af94",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de solicitações de viagem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "destination": {
                                                "type": "string",
                                                "example": "Nova Iorque"
                                            },
                                            "departure_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-08-01"
                                            },
                                            "return_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-08-10"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "solicitado"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Criar uma nova solicitação de viagem",
                "operationId": "c3b74eae7821f5c45d303c1648011c91",
                "requestBody": {
                    "description": "Dados da solicitação de viagem",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "destination",
                                    "departure_date",
                                    "return_date"
                                ],
                                "properties": {
                                    "destination": {
                                        "type": "string",
                                        "example": "Nova Iorque"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-01"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-08-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitação de viagem criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "destination": {
                                            "type": "string",
                                            "example": "Nova Iorque"
                                        },
                                        "departure_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-08-01"
                                        },
                                        "return_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-08-10"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "solicitado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trip-requests/{id}": {
            "get": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Obter uma solicitação de viagem específica",
                "operationId": "ef213aaade754916f8cae3a3ad2cf974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da solicitação de viagem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "destination": {
                                            "type": "string",
                                            "example": "Nova Iorque"
                                        },
                                        "departure_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-08-01"
                                        },
                                        "return_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-08-10"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "solicitado"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitação de viagem não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/trip-requests/{id}/status": {
            "patch": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Atualizar status da solicitação de viagem",
                "operationId": "e4eb043167ef4db42d6f415260129de0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da solicitação de viagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aprovado",
                                            "cancelado"
                                        ],
                                        "example": "aprovado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado com sucesso"
                    },
                    "403": {
                        "description": "Apenas administradores podem alterar o status"
                    },
                    "400": {
                        "description": "Viagens aprovadas não podem ser canceladas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}